requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
coloredlogs==15.0.1
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
executing==2.2.1
faiss-cpu==1.13.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
google-genai==1.52.0
googleapis-common-protos==1.72.0
groq==0.36.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
humanfriendly==10.0
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==8.6.3
jupyter-core==5.9.1
langchain==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-google-genai==3.2.0
langchain-groq==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.49
marshmallow==3.26.1
matplotlib-inline==0.2.1
mpmath==1.3.0
multidict==6.7.0
mypy-extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.5
onnxruntime==1.23.2
openai==2.8.1
opencv-python==4.11.0.86
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
pyzmq==27.1.0
rapidocr-onnxruntime==1.4.4
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.44
stack-data==0.6.3
structlog==25.5.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
